got playground proxy working
