+++ cerbos/sdk/client.py
@@ -1,10 +1,10 @@
 # Copyright 2021-2022 Zenauth Ltd.
 # SPDX-License-Identifier: Apache-2.0
 
-from cerbos.sdk._async.http import AsyncCerbosClient as AsyncCerbosClient
-from cerbos.sdk._async.http import AsyncPrincipalContext as AsyncPrincipalContext
-from cerbos.sdk._sync.http import CerbosClient as CerbosClient
-from cerbos.sdk._sync.http import PrincipalContext as PrincipalContext
+from cerbos.sdk._async._http import AsyncCerbosClient as AsyncCerbosClient
+from cerbos.sdk._async._http import AsyncPrincipalContext as AsyncPrincipalContext
+from cerbos.sdk._sync._http import CerbosClient as CerbosClient
+from cerbos.sdk._sync._http import PrincipalContext as PrincipalContext
 
 __all__ = [
     "AsyncCerbosClient",
