+++ cerbos/sdk/grpc/client.py
@@ -1,10 +1,10 @@
 # Copyright 2021-2023 Zenauth Ltd.
 # SPDX-License-Identifier: Apache-2.0
 
-from cerbos.sdk._async.grpc import AsyncCerbosClient as AsyncCerbosClient
-from cerbos.sdk._async.grpc import AsyncPrincipalContext as AsyncPrincipalContext
-from cerbos.sdk._sync.grpc import CerbosClient as CerbosClient
-from cerbos.sdk._sync.grpc import PrincipalContext as PrincipalContext
+from cerbos.sdk._async._grpc import AsyncCerbosClient as AsyncCerbosClient
+from cerbos.sdk._async._grpc import AsyncPrincipalContext as AsyncPrincipalContext
+from cerbos.sdk._sync._grpc import CerbosClient as CerbosClient
+from cerbos.sdk._sync._grpc import PrincipalContext as PrincipalContext
 
 __all__ = [
     "AsyncCerbosClient",
